Use FromIterator for conversions (#89)
Several collection conversions expected the `SymbolicByte` to be held in
a `Vec`, but there was no real reason for this expectation. This changes
the conversions from `From<Vec<...>>` to `FromIterator<...>`. The
calling code that current holds a `Vec` then calls with
`into_iter().collect()` instead of `.into()`.

This opens up the code to being able to use different collections
instead of being beholden to `Vec` due to the convenient conversion
routines.