Remove stored address spaces from memory (#92)
The emulator now handles all address lookup indirection. As a result the
memory module no longer needs to keep a record of the `AddressSpace` to
construct an address for `Load` and `Store` instructions.

As a result, the memory module will now write to any address provided,
even if the address space is not one of the known address spaces from
Sleigh. It is the responsibility of the caller to validate the address
and enforce any additional restrictions that are desired.

This additional validation is not currently performed except in the case
of an indirect address lookup. In those cases the target `Address` must
be constructed, including its `AddressSpace`, thereby ensuring the
`AddressSpace` is valid.