Completed migration to processor API (#98)
This completed the integration test migration to the `Processor` API
defined by source instead of the one defined for tests. In addition to
several changes to `Processor` required for completing this migration,
the introduction of new `mem` types was also required:

* `SymbolicMemoryReader` and `SymbolicMemoryWriter` traits which define
the `read` and `write` functions respectively. Implementing both of
these traits automatically implements `SymbolicMemory` via a blanket
implementation.
* `ExecutableMemory` is a new wrapper type which implements
`sla::LoadImage`.
* The `SymbolicMemoryWriter` API uses `IntoIterator` instead of
`ExactSizeIterator`. This simplifies the API on the caller side but
complicates it internally. Updates have not yet been made to callers to
take advantage of this change. #99