Added Sleigh trait and improved Sleigh errors (#126)
This update changes renamed `Sleigh` to `GhidraSleigh` and in favor of a
trait named `Sleigh`. This trait will help improve unit testing for the
processor API and also makes clear that the included implementation uses
the sleigh code from Ghidra.

As a part of this change, some APIs were changed to improve the trait
type signatures. This included a common `Disassembly` type for both
pcode and native disassemblies (previously called `DecodeResponse`).
Furthermore, the `String` errors were finally converted to a
`thiserror::Error` enumeration.

Note some APIs which cross FFI boundaries still have `String` types for
their errors. This is because they are ultimately serialized to their
`Display`. Including them in the `Error` enum implies they could
potentially be handled, but when we finally capture them on the Rust
side, they would be a String. Moreover, it would be difficult to
distinguish them from C++ exception errors (also captured as a
`String`). So to avoid giving the impression they can be handled as a
type, they are maintained as `String`.