Improvements to various APIs (#131)
* `VarnodeData` can now be tested for equality
* `Disassembly` data is now directly accessible instead of hidden behind
unnecessary accessors. The accessors created complications with data
borrowing, such as returning a reference to a specific
`PcodeInstruction`.
* `ConcretizationError` now records the symbolic bit directly instead of
in a combination of bit and byte index.
* `emulator::Error::SymbolicAddress` now encodes the address as
`SymbolicBitVec` instead of `Vec<SymbolicByte>`.
* `ControlFlow::ConditionalBranch` now includes the origin of the branch
condition
* `mem::Error::UnexpectedSymbolicData` was removed since it is no longer
used
* `mem::ExecutableMemory` was changed into a wrapper around a memory
reference. By not transferring ownership, the `ExecutableMemory` can be
constructed just prior to needing to load the instruction data. This
better supports alternative memory models.