Add z3 integration test, improvements to Aiger module (#134)
Added an integration test with `z3` solver. This test necessarily
exercised integration with the Aiger module that revealed opportunities
for enhancements.

1. The indexes for `And` gates were created in such a way that it made
it difficult to recreate them as constraints for the solver. This has
been updated to ensure a gate's index is always greater than both its
LHS and RHS.
2. `And` gate would always attempt to index its LHS and RHS components
even if it had already been indexed. This didn't result in incorrect
indexing, but the duplicated effort wasted a lot of time for large
circuits.