Performance improvements to several SymbolicBitVec APIs (#135)
This change updated the internals of `SymbolicBitVec` to use a
`VecDeque` instead of a `Vec`. This helped improve performance for
multiply and divide APIs. Also ported the shift implementations from
`SymbolicBitBuf`, which resulted in marked improvements to the shift
left and shift right APIs.

This change also removed some conversions from `SymbolicBitVec`.
Converting directly to a `SymbolicBit` was not used in practice, so this
was removed. This can still be achieved using `len() == 1` and `msb()`.

Additionally the `Deref` implementation into a slice has been removed,
since a `VecDeque` cannot make the same contiguous guarantees. The
common APIs used such as `len`, `is_empty`, and the `Index` trait impl
have been implemented directly on `SymbolicBitVec` instead.

Resolves #57 
Resolves #95