Use builder pattern for GhidraSleigh (#153)
The `GhidraSleigh` object required both a sleigh and processor
specification to be provided in the `initialize` function before further
functionality would ever succeed. This API is confusing since it does
not enforce the correct sequence. By applying the typestate builder
pattern here the compiler can enforce the necessary preconditions are
met.

This requires a version bump because the `new` method for construction
has been removed.