Fix coverage report showing as 0% in Github (#154)
It looks like there was a change at some point that resulted in the
generated `.profraw` files appearing in the crate roots instead of the
workspace root. Or it's possible that the prior crate organization in
this workspace was amenable to the default behavior but no longer is.

In either case the `.profraw` files appear in crate roots, so update the
merge to find all of them.