Added Pcode128 to pcode-ops crate (#155)
This struct was previously defined as a test fixture in the `pcode`
crate. However it serves as not just a useful example of an
implementation of `PcodeOps` but also has valid use as a simple concrete
value. This allows users of the `pcode-ops` crate to have a valid
implementation available immediately without needing to roll their own.