Add method for constructing PcodeOps value from PcodeOps bit (#156)
The methods `predicated_on` and `assert` in `PcodeOps` did not translate
to any actual pcode operation defined in the specification. However,
these functions were necessary functions for exploring pcode execution
contingent on branch conditions. This necessity was borne from the
inability to construct a pcode value from a branch condition,
specifically a value filled with that the branch condition bit.

This adds a new method `fill_bytes_with` to `PcodeOps` which creates a
value with the specified number of bytes filled with the desired
`PcodeOps::Bit`. The emulation now instead uses this more generic method
-- which is both simpler to understand and implement.