Add SymbolicBit evaluator  (#160)
This change adds an `Evaluator` to the `sym` crate that can evaluate a
`SymbolicBit` given `VariableAssignments`. This can be used to perform
concrete constraint guided symbolic execution by evaluating a
non-literal `SymbolicBit` with a concrete value when used in a branch
condition.

A constraint solver can then also be used to evaluate what different
input would result in taking the other path in the branch. An example of
this has been included in the integration test suite.