Replace C binary with Rust crate for pcode coverage integration test (#164)
This change replaces the binary generated from C code with a Rust crate
capable of producing the binary. Currently the compiler version for this
binary is not pinned, which leaves it open to compiler changes possibly
inducing different amounts of coverage. This can be mitigated with
converting more of the Rust code into assembly code.

This change also adds some additional debugging improvements to some
`libsla` objects.