Remove `TryInto<u64>` requirement from `PcodeOps` (#166)
This requirement was largely required for the emulator to convert values
into offsets for indirect branches. However all the machinery is already
in place to accomplish this without requiring `TryInto<u64>`. It also
strangely specialized the requirement around `u64` but did not have
support for other values, leading to awkward usage in those cases.

The direct requirement has been removed and replaced with a wrapper
`PcodeValue` for integral type conversions. This is used for `u64` in
the emulator (replacing the direct `TryInto<u64>` conversion). It also
simplified the x86 emulator integration test conversion logic which
needed to convert to `u128`.