Moved additional Linux support from integration test into source (#174)
This change introduces the Linux emulator and the x86 processor handler
into the source tree. Both of these previously existed in the
integration test but are crucial for exercising this architecture and
kernel.

The interface for the processor handlers trait was updated to remove the
generic requirement on the memory, instead tying this to the functions
specifically. This eliminates a difficulty with maintaining a generic
handler. The integration test did not encounter this difficulty due to
specializing on the implementation used in the test.