Add x86 emulator with Linux kernel support (#176)
This updates a recent change which added an emulator specifically with
Linux support. The issue there is the Linux kernel is largely
architecture agnostic -- `mmap` means the same thing on all
architectures. The kernel _does_ have architecture-specific logic --
importantly, which registers are used to pass arguments to `syscall`.

With this update there are now architecture-specific emulators
(currently only x86) which know which Pcode instruction represents a
syscall, and to invoke the kernel to handle it. There is a default
`NoKernel` implementation which can be used to avoid modeling the
kernel.

In order to support modeling an x86 `Processor`, the corresponding
handler logic that previously lived in the integration test has been
lifted to the source tree.

Finally, this change includes a new `ControlFlow::Halt` variant to
handle the exit status. This was previously implemented with an error
but was clumsy to decode, since it required downcasting an
`Error::DependencyError`. This new control flow is never returned by the
core emulator logic.