Add Sleigh API to get all registers with names (#177)
A new `registers` API has been added to `Sleigh` to retrieve a sorted
map of registers to names. The map is sorted such that `VarnodeData`
with the lowest `Address` offset is returned first. In the event of
multiple overlapping registers with identical offsets (e.g. `RAX` and
`EAX` in x86), the `VarnodeData` with the _largest_ size is returned
first.