Upgrade to Ghidra 11.4 (#190)
This upgrades libsla to Ghidra to 11.4. Most importantly the sleigh
compiler now produces a binary file instead of an XML file. The sleigh
specification expected by the underlying sleigh impl is now required to
be `<sleigh>path/to/binspec.sla</sleigh>`.

This change also moves the `sys` module into its own `libsla-sys` crate
for easier sharing. Linking the sleigh-compiler could fail with
duplicate objects if its version did not exactly match the liblsa
version. Once published the sleigh compiler will be updated to remove
its dependency on Ghidra and instead depend on the shared sys crate.

Finally the x86 emulator encountered a new unlock instruction in the
Linux integration test that did not exist prior. Handling for this
instruction was added to resolve the failing test.