Remove libsla-sys crate (#192)
This crate has been moved to its own Github repository at
https://github.com/mnemonikr/libsla-sys to simplify testing. By keeping
it locally, it would force the local version when building and fail to
pull in the latest sleigh-compiler which depended on the crate version.
This is due to the fact that only one version of a sys crate (enforced
via the `links` attribute in `Cargo.toml`) can be used in a dependency
tree.

The Ghidra repository has been moved to the integration tests crate
since it is only used internally during tests for its processor
specification files (slaspec and pspec).