Simplify Processor, add BranchingProcessor, and processor module cleanup (#193)
This changes the `Processor` object to return an error in the event of a
symbolic branch. A new method `step_branch` has been added for a caller
to provide the concrete evaluation of the condition of the symbolic
branch. This way if the caller is not attempting to explore multiple
paths in parallel, the caller no longer needs to juggle processors.

A `BranchingProcessor` version has been added that largely retains the
original behavior of the `Processor` branching on step. This uses the
new `Processor` internally. The key difference the new processor
returned from its `step` returns the processor that _takes_ the branch
instead of the one that does not.

The `ProcessorManager` has been removed since it did not provide much
additional functionality over a standard `Vec`.

Finally there was some internal cleanup on the processor module, such as
adding a `Display` to the `ProcessorState` and an accessor for the pcode
execution index.

Resolves #182