Expanded Linux kernel model to support multiple architectures, AArch64 demo (#196)
This adds support to the Linux kernel model to support additional
architectures. Included AArch64 model with sufficient syscalls to
execute hello world. Added this AArch64 example to the integration
tests.

This assumes the build is being performed on an x86-64 platform which
necessitates cross-compiling to AArch64.

The Linux kernel model still defaults to x86-64 if additional
configuration has not been provided. This will later be removed and be
required to provide the architecture details.