Changes to Evaluator in sym crate (#197)
This makes a number of changes to the `Evaluator` in the `sym` crate.
The API has been changed to return an `Evaluation` which includes
additional details. The evaluator will no longer panic if the response
is not a concrete value; in this case it will be `None` instead.

The evaluation also tracks which variables were used to perform the
evaluation, as well as any unassigned variables encountered during the
evaluation. The latter is only populated when the response is not
concrete.

Finally the evaluator cache has been removed due to a known issue. This
optimization is not necessary to include in the core evaluator.

Resolves #179