Add support for decoding raw sla format (#208)
This adds `GhidraSleighBuilder::sla_decoder` which can be used when not
loading the sla spec from the document store. This method can be used to
specify the decoding format, either `sla` (the normal format used by
Ghidra) or `raw`, a modification of the standard format that omits the
header and does not compress the data. The `sla` format is still the
default format.