Enable additional SymbolicBit reductions (#231)
The `SymbolicBit` internally has pointers, though the bit itself is
_not_ a pointer. This meant `x.clone() ^ x` would in general not
evaluate to false, because the cloned value is a distinct object. To
remedy this the `BitAnd` operator (the fundamental operation underlying
all others) has been updated to check for `x & x` as well as `x & !x`.
This latter case was previously only handled if `x` was a variable.

Confirmed `x ^ x` now evaluates to `false` even when `x` is a complex
object.

This required bumping the MSRV to 1.88 that's when certain patterns of
`if-let` chains were stabilized:
https://github.com/rust-lang/rust/issues/53667